renamed package to PynkTromboneGymnasium
