+++ pynktrombonegymnasium/spectrogram.py
